Update video-indexer.js
Added "waitForProcessing", which allows the caller to simply wait for the file to be processed while the library itself polls the service. The promise is resolved when the file is 100% processed. This also enables concurrent processing of multiple files easily from a client library. Additionally, I have a few of the promises bind to the current context so they're a little easier to work with. I also added a helper method to download VTT files and chain that with the retrieval of the VTT URL.